# Improve serialization perf