--- a/bench/bench-shell-bindings.js
+++ b/bench/bench-shell-bindings.js
@@ -1,3 +1,16 @@
+if (typeof load !== "function") {
+  var fs = require("fs");
+  var vm = require("vm");
+  load = function(file) {
+    var src = fs.readFileSync(file, "utf8");
+    vm.runInThisContext(src);
+  };
+}
+
+if (typeof print !== "function") {
+  print = console.log.bind(console);
+}
+
 load("./scalajs-runtime-sourcemap.js");
 load("./stats.js");
 load("../dist/source-map.js");