--- a/lib/source-map/array-set.js
+++ b/lib/source-map/array-set.js
@@ -49,13 +49,14 @@ define(function (require, exports, module) {
    * @param String aStr
    */
   ArraySet.prototype.add = function ArraySet_add(aStr, aAllowDuplicates) {
-    var isDuplicate = this.has(aStr);
+    var sStr = util.toSetString(aStr);
+    var isDuplicate = this._set.hasOwnProperty(sStr);
     var idx = this._array.length;
     if (!isDuplicate || aAllowDuplicates) {
       this._array.push(aStr);
     }
     if (!isDuplicate) {
-      this._set[util.toSetString(aStr)] = idx;
+      this._set[sStr] = idx;
     }
   };
 
@@ -65,8 +66,8 @@ define(function (require, exports, module) {
    * @param String aStr
    */
   ArraySet.prototype.has = function ArraySet_has(aStr) {
-    return Object.prototype.hasOwnProperty.call(this._set,
-                                                util.toSetString(aStr));
+    var sStr = util.toSetString(aStr);
+    return this._set.hasOwnProperty(sStr);
   };
 
   /**
@@ -75,8 +76,9 @@ define(function (require, exports, module) {
    * @param String aStr
    */
   ArraySet.prototype.indexOf = function ArraySet_indexOf(aStr) {
-    if (this.has(aStr)) {
-      return this._set[util.toSetString(aStr)];
+    var sStr = util.toSetString(aStr);
+    if (this._set.hasOwnProperty(sStr)) {
+      return this._set[sStr];
     }
     throw new Error('"' + aStr + '" is not in the set.');
   };