--- a/lib/source-map/source-map-generator.js
+++ b/lib/source-map/source-map-generator.js
@@ -295,6 +295,8 @@ define(function (require, exports, module) {
       var previousSource = 0;
       var result = '';
       var mapping;
+      var nameIdx;
+      var sourceIdx;
 
       var mappings = this._mappings.toArray();
       for (var i = 0, len = mappings.length; i < len; i++) {
@@ -321,9 +323,9 @@ define(function (require, exports, module) {
         previousGeneratedColumn = mapping.generatedColumn;
 
         if (mapping.source != null) {
-          result += base64VLQ.encode(this._sources.indexOf(mapping.source)
-                                     - previousSource);
-          previousSource = this._sources.indexOf(mapping.source);
+          sourceIdx = this._sources.indexOf(mapping.source);
+          result += base64VLQ.encode(sourceIdx - previousSource);
+          previousSource = sourceIdx;
 
           // lines are stored 0-based in SourceMap spec version 3
           result += base64VLQ.encode(mapping.originalLine - 1
@@ -335,9 +337,9 @@ define(function (require, exports, module) {
           previousOriginalColumn = mapping.originalColumn;
 
           if (mapping.name != null) {
-            result += base64VLQ.encode(this._names.indexOf(mapping.name)
-                                       - previousName);
-            previousName = this._names.indexOf(mapping.name);
+            nameIdx = this._names.indexOf(mapping.name);
+            result += base64VLQ.encode(nameIdx - previousName);
+            previousName = nameIdx;
           }
         }
       }