@fitzgen 2015-08-25
Looks great! Thanks! I'm just going to run the benchmarks locally, real quick.